fix 71-minute timestamp wrap: emit 64-bit ticks_us from firmware
fw_ticks_us was truncated to uint32 (wraps at ~71.6 min), causing
elapsed_s to go negative and breaking step detection. Now all three
timestamp emission points (D lines, P clock sync, G anchor) use the
full uint64 from time_us_64(). Python's int() handles 64-bit natively.

Also raised MAX_FREQ from 300 to 1000 Hz in the TUI.

+++ firmware-c/src/switching.h
@@ -59,7 +59,7 @@ uint32_t switching_get_period_us(void);
 //   - the firmware ticks_us at the moment state 0 went on the pins is
 //     captured into *out_anchor_ticks_us (used by the OK G reply for
 //     Pi-side clock anchoring; see main.py:_switching_start docstring).
-bool switching_start(uint32_t *out_anchor_ticks_us);
+bool switching_start(uint64_t *out_anchor_ticks_us);
 
 // Halt periodic switching. All FETs go off. Idempotent.
 void switching_halt(void);
